[chore] 집들이 API REST DOCS 명세
